Fix test code, add more tests
